requirements.txt
Flask==2.3.2
qrcode==7.4.2
Pillow==10.0.0
Werkzeug==2.3.6